Lab_work_6/user_control_save_in_kafka/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
kafka-python==2.0.2
passlib==1.7.4
pydantic==2.10.6